move stuff into the case data provider class
